requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
drf-spectacular==0.26.2
python-decouple==3.8
Pillow>=10.0.0
django-filter==23.1